move FragmentProcessingException to fragmetns api
